requirements.txt
Django==1.10.2
argparse==1.2.1
dj-database-url==0.4.1
django-cors-headers==1.2.2
djangorestframework==3.4.7
gunicorn==19.4.5
psycopg2==2.6.2
requests==2.11.1
whitenoise==2.0.6
wsgiref==0.1.2
Pillow==3.4.2
